refactor: extract MarketTableHeader to separate file
Move MarketTableHeader component and shared column width constants
to components/MarketTableHeader.tsx for better code organization.
