+++ packages/kit/src/views/UniversalSearch/pages/UniversalSearch.tsx
@@ -18,7 +18,6 @@ import {
   View,
   XStack,
   YStack,
-  useMedia,
 } from '@onekeyhq/components';
 import { DiscoveryBrowserProviderMirror } from '@onekeyhq/kit/src/views/Discovery/components/DiscoveryBrowserProviderMirror';
 import { EJotaiContextStoreNames } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
@@ -61,6 +60,7 @@ import {
   UniversalSearchV2MarketTokenItem,
 } from '../components/SearchResultItems';
 
+import { MarketTableHeader } from '../components/MarketTableHeader';
 import { RecentSearched } from './components/RecentSearched';
 import { UniversalSearchProviderMirror } from './UniversalSearchProviderMirror';
 
@@ -128,60 +128,6 @@ function ListEmptyComponent() {
 }
 
 const isMarketSection = (tabIndex: number) => tabIndex === 2;
-const MARKET_NAME_COLUMN_WIDTH = 160;
-const MARKET_DATA_COLUMN_WIDTH = '33.3333%';
-
-function MarketTableHeader() {
-  const intl = useIntl();
-  const { gtMd } = useMedia();
-  return (
-    <XStack alignSelf="stretch" mx="$2" px="$3" py="$1.5" gap="$3" ai="center">
-      {/* Matches row: star($8) + gap($1) + icon($8) + gap($2) + name */}
-      <XStack w={MARKET_NAME_COLUMN_WIDTH} gap="$1" ai="center" flexShrink={0}>
-        <XStack w="$8" ai="center" jc="center">
-          <SizableText size="$bodySm" color="$textSubdued">
-            #
-          </SizableText>
-        </XStack>
-        <SizableText size="$bodySm" color="$textSubdued" flex={1}>
-          {intl.formatMessage({ id: ETranslations.global_name }).toUpperCase()}
-        </SizableText>
-      </XStack>
-      <XStack flex={1} minWidth={0}>
-        <XStack
-          w={gtMd ? MARKET_DATA_COLUMN_WIDTH : undefined}
-          flex={gtMd ? undefined : 1}
-          jc="flex-end"
-        >
-          <SizableText size="$bodySm" color="$textSubdued" textAlign="right">
-            {intl
-              .formatMessage({ id: ETranslations.global_price })
-              .toUpperCase()}{' '}
-            / 24H
-          </SizableText>
-        </XStack>
-        {gtMd ? (
-          <XStack w={MARKET_DATA_COLUMN_WIDTH} jc="flex-end">
-            <SizableText size="$bodySm" color="$textSubdued" textAlign="right">
-              {intl
-                .formatMessage({ id: ETranslations.global_liquidity })
-                .toUpperCase()}
-            </SizableText>
-          </XStack>
-        ) : null}
-        {gtMd ? (
-          <XStack w={MARKET_DATA_COLUMN_WIDTH} jc="flex-end">
-            <SizableText size="$bodySm" color="$textSubdued" textAlign="right">
-              {intl
-                .formatMessage({ id: ETranslations.dexmarket_turnover })
-                .toUpperCase()}
-            </SizableText>
-          </XStack>
-        ) : null}
-      </XStack>
-    </XStack>
-  );
-}
 
 export function UniversalSearch({
   filterTypes,
